% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ANA.R
\name{ANA_dados}
\alias{ANA_dados}
\title{Puxar dados para estações da ANA}
\usage{
ANA_dados(
  data_inicio = "01/01/2021",
  data_fim = format(Sys.Date(), "\%d/\%m/\%Y"),
  codigo_estacao = c(60478482, 60479280)
)
}
\arguments{
\item{data_inicio}{primeira data que se quer os dados}

\item{data_fim}{última data que se quer os dados. O *default* é a data atual.}

\item{codigo_estacao}{pode ser um valor único de uma estação ou um vetor
contendo diversas estações.}
}
\value{
dataframe
}
\description{
Puxar dados para estações da ANA
}
\examples{
teste_ANA_dados <- ANA_dados()
}
