% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_ANA.R
\name{ANA_info}
\alias{ANA_info}
\title{Puxar informações sobre as estações da ANA dentro do DF}
\usage{
ANA_info(estacao_codigo = "", origem_codigo = "", buffer_DF = 10000)
}
\arguments{
\item{estacao_codigo}{se a estação está ativa (0) ou em manutenção (1).
O *default* é puxar os dois tipos}

\item{origem_codigo}{de quem é o dado. O *default* é puxar todos.
Os valores podem ser:
0-Todas, 1-ANA/INPE, 2-ANA/SIVAM, 3-RES_CONJ_03, 4-CotaOnline, 5-Projetos Especiais.}

\item{buffer_DF}{tamanho do buffer em volta do DF. O *default* é de 10km.
Para não utilizar um *buffer* basta utilizar buffer_DF = 0.}
}
\value{
sf dataframe
}
\description{
Puxar informações sobre as estações da ANA dentro do DF
}
\examples{
teste_ANA_info <- ANA_info()
ANA_info_sem_buffer <- ANA_info(buffer_DF = 0)
}
